Simplify, allow for unencrypted key

+++ jobs/ssh-init.sh
@@ -16,15 +16,9 @@
 # NOTE:
 # The use of a FIFO and cat <<EOF is to prevent the passphrase from ever
 # being stored in a file or appearing in the process list; otherwise this
-# script might be a lot shorter.
+# script might be a bit shorter.
 
-if [ -z "$BOT_SSH_PHRASE" ]
-then
-    MESSAGE="BOT_SSH_PHRASE not set, see conf/jobs/ssh-init.yaml"
-    Log "Error" "$MESSAGE"
-    echo "$MESSAGE" >&2
-    exit 1
-fi
+source $GOPHER_INSTALLDIR/lib/gopherbot_v1.sh
 
 if [ -z "$HOME" ]
 then
@@ -34,34 +28,45 @@ then
     exit 1
 fi
 
-source $GOPHER_INSTALLDIR/lib/gopherbot_v1.sh
+SSH_KEY=${KEYNAME:-id_rsa}
+SSH_KEY_PATH="$HOME/.ssh/$SSH_KEY"
 
-STMP=$(mktemp -d -p $HOME ssh-agent-XXXXXXX)
-
-cat <<EOF >$STMP/askpass.sh
+if grep -q ENCRYPTED $SSH_KEY_PATH
+then
+    ENCRYPTED=true
+    if [ -z "$BOT_SSH_PHRASE" ]
+    then
+        MESSAGE="BOT_SSH_PHRASE not set, see conf/jobs/ssh-init.yaml"
+        Log "Error" "$MESSAGE"
+        echo "$MESSAGE" >&2
+        exit 1
+    fi
+    STMP=$(mktemp -d -p $HOME ssh-agent-XXXXXXX)
+    cat <<EOF >$STMP/askpass.sh
 #!/bin/bash
 cat $STMP/sfifo
 EOF
-chmod +x $STMP/askpass.sh
+    chmod +x $STMP/askpass.sh
 
-mkfifo $STMP/sfifo
-mkfifo $STMP/afifo
-cat <<EOF >$STMP/sfifo &
+    mkfifo $STMP/sfifo
+    cat <<EOF >$STMP/sfifo &
 $BOT_SSH_PHRASE
 EOF
+    export SSH_ASKPASS=$STMP/askpass.sh
+    export DISPLAY=""
+fi
 
-ssh-agent >$STMP/afifo &
-eval `cat $STMP/afifo`
+eval `ssh-agent`
 
-export SSH_ASKPASS=$STMP/askpass.sh
-export DISPLAY=""
-SSH_KEY=${KEYNAME:-id_rsa}
 ssh-add $HOME/.ssh/$SSH_KEY < /dev/null
 
 # Make agent available to other tasks in the pipeline
 SetParameter SSH_AUTH_SOCK $SSH_AUTH_SOCK
 SetParameter SSH_AGENT_PID $SSH_AGENT_PID
 
-rm -rf $STMP
+if [ -n "$ENCRYPTED" ]
+then
+    rm -rf $STMP
+fi
 
 exit 0
